update AWBM-Calibration.py
Starting to migrate from the old implementation of the AWBM(lists and appending)  into a dataframe containing all the variables that is appended.

From that past data processing experience, the lists seemed really slow to work with - particularly when dealing with combining lists into a DF for export or having to fix data types and formatting.