update AWBM-Calibration.py
Added the ability to easily animate the plots produced.

From this, it looks like Cavg alone may not be enough to calibrate effectively, but I need to test the model against data that suits the initial condition assumptions (empty or full soil stores).

Going to start working on the rainfall analysis script to easily identify the longest continuous dry and wet periods from the SILO data.